#4 - Adapting the Light theme to the latest progress made with the Dark theme
Also resolves #1

While refactoring the theme, it became apparent that due to not using as many custom, non-standard colors in the original Light theme iteration, we lost a bit of contrast overall. The topic of contrast will be handled in a separate issue.

In addition, properties that were deprecated were cleaned up from the config.